feat(tape,experiment,memory-graph): finish Crucible tape integration + experiment-API hardening + edges self-heal
Lands and corrects the large multi-agent uncommitted changeset (Crucible's
context-tape paging subsystem + P9 benchmark; liblevenshtein-requested generic
experiment-API improvements) and fixes a pre-existing production bug. Files are
deeply intertwined (shared migrations.rs / lib.rs / server.rs / config.rs), so
this is one coherent commit. Full analysis: docs/reviews/uncommitted-cowboy-changes-2026-06-20.md.

memory-graph (graph_neighbors outage):
- migrations.rs: matviews_present + ensure_edges_only; the startup guard and the
  hash gate both check BOTH matviews and cheaply repair a missing memory_unified_
  edges (existence dominates the hash gate; non-fatal on the critical path); edges
  built BEFORE the nodes HNSW so the missing-edges window collapses. Self-heals on
  restart. Test: memory_unified_edges_self_heal.rs.

tape integration (was high-craft but UNWIRED):
- v53 migration: relax the working_set FK so an RLM tree's synthetic session_key
  ("rlm:{root_task_id}") can persist (the reason the engine had no production
  caller), add working_set_pages.content (durable scratch bytes), index tree_path,
  preserve CSM cascade via a trigger.
- engine: single clock authority (atomic bump_clock; save_config no longer
  overwrites the clock); byte carriage on ResidentPage (Some only for Scratch);
  real dirty write-back; implemented TTL eviction (EvictReason::Ttl); removed the
  throwaway DynamicDawg from LRU/TTL/FIFO; admit_scratch (the RAM↔DB unification);
  EvictionPolicyEngine: Send + Sync (handler futures must be Send).
- store: atomic save_working_set (one tx); content load/save; rehydrate_store_from_pages.
- tape_put routes through admit_scratch (residency-tracked + durable); checkpoint
  resume rehydrates the TapeStore from working_set_pages; reaper cron + drop_tree
  at root-frame completion; wired the 3 dead [tape] config fields.
- rlm: namespaced accumulator slots accum/<parent_task_id>/<ord> (fixes the depth>1
  concurrent Store-stitch collision); drop_tree at the root frame.
- registry: drop the DashMap shard guard before locking the inner Mutex.
- real_data_plane: batch the N+1 chunk-id lookups; correct the "corpus READ-ONLY"
  docs (memory_observations is gated-promotion-writable).
- tape_repl: thread the real caller identity (extract_caller), lowercase
  black_box_roles, fail closed on unknown — the wire gate no longer fails open.

experiment-API hardening (generic; firewalled from the tracker):
- v52 migration: ExperimentRunStatus closed vocab + status CHECK, the immutable
  experiment_run_status_audit trail, a tamper-evident samples_digest, and the
  experiment_paired_binary 2x2 table.
- McNemar + exact-binomial stats (inference.rs); experiment_decide consumes only
  decision-usable runs (load_experiment_samples gated on usable status).
- tools: experiment_record_paired_binary_counts (server-computed McNemar verdict),
  experiment_finalize_run, experiment_set_run_status (audited; post-decision
  invalidation re-opens the decision), experiment_record_measurement_from_artifact
  (path-traversal-safe CSV/JSONL), and experiment_get audit-visibility expansion.

P9 (src/experiment/context_tape.rs): kept as-is (verified) — a pre-registered,
server-computed, default-OFF memory-promotion benchmark that does not touch the
tracker and fabricates no data.

Also: the tape DB-test pool helpers now use an isolated PGMCP_TEST_DATABASE_URL +
run migrations (never the live DB); incidental clippy/fmt fixes in two unrelated
test files.

+++ pgmcp-testing/tests/topic_apps3.rs
@@ -83,7 +83,7 @@ async fn topic_ownership_forecast_flags_single_owner() {
              ON CONFLICT (file_id, chunk_index) DO UPDATE SET blame_author='alice' RETURNING id",
         )
         .bind(file_id)
-        .bind(i as i32)
+        .bind(i)
         .fetch_one(&pool)
         .await
         .expect("chunk");
